Fixing Bug on RandomUniformValueSignal
In most of the cases, this class should provide the value for the delay.
During the initialization, for some reason, the first collected value
was zero, messing up the delay. This commit sets the lower limit of
the initial value to y_min, fixing the problem.

Besides that, during debug, some classes were created in order to
better organize the models. The created classes were DataCollect
and Transition.